% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_operation.R
\name{addition}
\alias{addition}
\title{Addition of a POSIXlt date to a value in second}
\usage{
addition(date, sec)
}
\arguments{
\item{date}{POSIXlt date}

\item{sec}{value in second}
}
\description{
Addition of a POSIXlt date to a value in second
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
addition(date=res,sec=487654)

}
\keyword{timeManip}

