% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition_char_char.R
\name{addition_char_char}
\alias{addition_char_char}
\title{Addition of a char date to a char value at a specific unit}
\usage{
addition_char_char(char, date_chr, v, unit)
}
\arguments{
\item{char}{"YYYY" or "YYYYmm" or "YYYYmmdd" or "YYYYmmddHH" or "YYYYmmddHHMM" or "YYYYmmddHHMMSS"}

\item{date_chr}{date in character}

\item{v}{value}

\item{unit}{choice between daily","hourly","minute","second"}
}
\description{
Addition of a char date to a char value at a specific unit
}
\examples{
addition_char_char(char="YYYYmmddHH",date_chr="2013060205",v=3,unit="hourly")

}
\keyword{timeManip}

