% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_operation.R
\name{addition_nonsec}
\alias{addition_nonsec}
\title{Addition of a POSIXlt date to a value not in second}
\usage{
addition_nonsec(date, timeResolution, v)
}
\arguments{
\item{date}{POSIXlt date}

\item{timeResolution}{"daily","hourly","minute","second"}

\item{v}{value}
}
\description{
Addition of a POSIXlt date to a value not in second
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
addition_nonsec(date=res,"hourly",v=10)

}
\keyword{timeManip}
