% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition_char_char.R
\name{addition_plt_sec}
\alias{addition_plt_sec}
\title{Addition of a POSIXlt date to a value in second}
\usage{
addition_plt_sec(date, sec)
}
\arguments{
\item{date}{POSIXlt date}

\item{sec}{value in second}
}
\description{
Addition of a POSIXlt date to a value in second
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
addition_plt_sec(date=res,sec=487654)

}
\keyword{timeManip}

