% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2POSIXct.R
\name{char2POSIXct}
\alias{char2POSIXct}
\title{char2POSIXct}
\usage{
char2POSIXct(char, date_chr)
}
\arguments{
\item{char}{"YYYY" or "YYYYmm" or "YYYYmmdd" or "YYYYmmddHH" or "YYYYmmddHHMM" or "YYYYmmddHHMMSS"}

\item{date_chr}{date in character}
}
\value{
The output is a POSIXct value
}
\description{
char2POSIXct
}
\examples{
res <- char2POSIXct("YYYYmmddHH","2015010100")
res <- char2POSIXct("YYYYmm","198002")
}
\keyword{timeManip}

