% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2POSIXlt.R
\name{char2POSIXlt}
\alias{char2POSIXlt}
\title{char2POSIXlt}
\usage{
char2POSIXlt(char, date_chr)
}
\arguments{
\item{char}{"YYYY" or "YYYYmm" or "YYYYmmdd" or "YYYYmmddHH" or "YYYYmmddHHMM" or "YYYYmmddHHMMSS"}

\item{date_chr}{date in character}
}
\value{
The output is a POSIXct value
}
\description{
char2POSIXlt
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
res <- char2POSIXlt("YYYYmm","198002")
}
\keyword{timeManip}

