% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contain.R
\name{contain}
\alias{contain}
\title{contain}
\usage{
contain(ts, ts_sub)
}
\arguments{
\item{ts}{POSIXlt timeserie}

\item{ts_sub}{POSIXlt sub timeserie}
}
\description{
contain
}
\examples{
\dontrun{
contain()
}
}
\keyword{timeManip}
