% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFormat.R
\name{convertFormat}
\alias{convertFormat}
\title{Convert the format of a POSIXct date}
\usage{
convertFormat(date, formatTo)
}
\arguments{
\item{date}{format POSIXct}

\item{format}{conversion into a vector "YYMMDDHH" or a matrix "YY,MM,DD,HH"}
}
\description{
Convert the format of a POSIXct date
}
\examples{
res <- convertFormat(strptime("2016032701","\%Y\%m\%d\%H\%",tz="GMT"),"YY,MM,DD,HH")
res <- convertFormat(strptime("2016032701","\%Y\%m\%d\%H\%",tz="GMT"),"YYMMDDHH")
}
\keyword{timeManip}

