% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeSerie.R
\name{dateTimeSerie}
\alias{dateTimeSerie}
\title{Make a timeserie of dates}
\usage{
dateTimeSerie(timeResolution, fromPeriod, toPeriod)
}
\arguments{
\item{timeResolution}{choice between "daily","hourly","minute","second"}

\item{fromPeriod}{"YYYYmmddHHMMSS"}

\item{toPeriod}{"YYYYmmddHHMMSS"}
}
\value{
The output is a list of a timeserie of POSIXct and the number of steps
}
\description{
The fuction \code{dateTimeSerie()} builds the timeserie of date
}
\examples{
res <- dateTimeSerie("hourly","20150101","20150103")
head(res)
res <- dateTimeSerie("hourly","19800228","19800301")
head(res)
}
\keyword{timeManip}

