% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{df2time}
\alias{df2time}
\title{df2time}
\usage{
df2time(precision, df)
}
\arguments{
\item{precision}{choice between "daily","hourly","minute","second",...}

\item{df}{date.frame of the columns year, months,...}
}
\value{
The output is a POSIXlt value
}
\description{
df2time
}
\examples{
df <- data.frame(year=rep(2015,5),month=c(1,2,3,4,5),day=c(1,1,1,1,1),hour=c(1,2,3,4,5))
df2time(precision="hourly",df)
}
\keyword{timeManip}
