% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{difference}
\alias{difference}
\title{Difference between two dates}
\usage{
difference(date1, date2)
}
\arguments{
\item{date1}{date POSIXct}

\item{date2}{date POSIXct}
}
\description{
Difference in seconds between two dates
}
\examples{
\dontrun{
difference()
}
}
\keyword{timeManip}

