% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{getP}
\alias{getP}
\title{Get the part of the POSIXlt date you want}
\usage{
getP(date, part)
}
\arguments{
\item{date}{POSIXlt}

\item{part}{"year" "month", "day", "hour", "minute", "second"}
}
\description{
Get the part of the POSIXlt date you want
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
getP(res,"month")
}
\keyword{timeManip}

