% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{idist}
\alias{idist}
\title{Get the idist}
\usage{
idist(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the idist
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
idist(res)
}
\keyword{timeManip}

