% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{indice_subdate}
\alias{indice_subdate}
\title{indice_subdate}
\usage{
indice_subdate(date, d)
}
\arguments{
\item{date}{timeManip object}

\item{d}{sub timeManip object}
}
\description{
indice_subdate
}
\examples{
a <- timeManip(fromPeriod="2013060205",toPeriod="2013060605",
               timeResolution="hourly",precision="hourly")
b <- timeManip(fromPeriod="2013060305",toPeriod="2013060505",
               timeResolution="hourly",precision="hourly")
res <- indice_subdate(a,b)
str(res)
}
\keyword{timeManip}

