% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{isdst}
\alias{isdst}
\title{Get the isdst}
\usage{
isdst(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the isdst
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
isdst(res)
}
\keyword{timeManip}
