% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Keep the part of the date you want}
\usage{
keep(fromPeriod = "20160420", toPeriod = "20160630",
  timeResolution = "daily", part = "month")
}
\arguments{
\item{fromPeriod}{"YYYYmmddHHMMSS"}

\item{toPeriod}{"YYYYmmddHHMMSS"}

\item{timeResolution}{choice between "daily","hourly","minute","second"}

\item{part}{either "year" "month", or "day"}
}
\description{
The function \code{keep()} extract the part of the date you want
}
\examples{
keep()
}
\keyword{timeManip}

