% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{minute}
\alias{minute}
\title{Get the minute}
\usage{
minute(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the minute
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
minute(res)
}
\keyword{timeManip}

