% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{month}
\alias{month}
\title{Get the month}
\usage{
month(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the month
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
month(res)
}
\keyword{timeManip}

