% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXlt2char.R
\name{POSIXlt2char}
\alias{POSIXlt2char}
\title{POSIXlt2char}
\usage{
POSIXlt2char(date, char)
}
\arguments{
\item{date}{POSIXlt date}

\item{char}{"YYYY" or "YYYYmm" or "YYYYmmdd" or "YYYYmmddHH" or "YYYYmmddHHMM" or "YYYYmmddHHMMSS"}
}
\description{
POSIXlt2char
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
POSIXlt2char(res,"YYYYmm")
}
\keyword{timeManip}

