% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{getFromNamespace}
\alias{reexports}
\alias{str}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{utils}{\code{\link[utils]{getFromNamespace}}, \code{\link[utils]{str}}}
}}

