% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{standard}
\alias{standard}
\title{standard}
\usage{
standard(precision, date_chr)
}
\arguments{
\item{precision}{choice between "daily","hourly","minute","second",...}

\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
standard
}
\examples{
res <- standard("hourly","2015010100")
res <- standard("monthly","198002")
}
\keyword{timeManip}
