% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{standard2}
\alias{standard2}
\title{standard2}
\usage{
standard2(precision, date_chr)
}
\arguments{
\item{precision}{choice between "daily","hourly","minute","second",...}

\item{date_chr}{date in character as "2015/01/12 06:00"}
}
\value{
The output is a POSIXlt value
}
\description{
standard2
}
\examples{
res <- standard2("hourly","2015/01/01 00")
res <- standard2("monthly","1980/02")
res <- standard2("monthly","1980/2/1 06:12")
}
\keyword{timeManip}
