% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_timeserie.R
\name{timeserie}
\alias{timeserie}
\title{Make a timeserie of dates}
\usage{
timeserie(timeResolution, fromPeriod, toPeriod, precision = NULL)
}
\arguments{
\item{timeResolution}{"daily","hourly","minute","second"}

\item{fromPeriod}{date char "YYYYmmddH..."}

\item{toPeriod}{date char "YYYYmmdd..."}

\item{precision}{"daily","hourly","minute","second",...}
}
\value{
The output is a list of a timeserie of POSIXct and the number of steps
}
\description{
The fuction \code{dateTimeSerie()} builds the timeserie of date
}
\examples{
res <- timeserie(timeResolution = "hourly",fromPeriod="2015010100",toPeriod="2015010322")
head(res)
res <- timeserie("daily","1980022802","1980030102","hourly")
head(res)
}
\keyword{timeManip}

