% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{tzone}
\alias{tzone}
\title{Get the tzone}
\usage{
tzone(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the tzone
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
tzone(res)
}
\keyword{timeManip}

