% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{wday}
\alias{wday}
\title{Get the wday}
\usage{
wday(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the wday
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
wday(res)
}
\keyword{timeManip}
