% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{yday}
\alias{yday}
\title{Get the yday}
\usage{
yday(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the yday
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
yday(res)
}
\keyword{timeManip}

