% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{year}
\alias{year}
\title{Get the year}
\usage{
year(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the year
}
\examples{
res <- char2POSIXlt("YYYYmmddHH","2015010100")
year(res)
}
\keyword{timeManip}

