% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYY}
\alias{YYYY}
\title{Get YYYY}
\usage{
YYYY(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year as character
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYY(res)
}
\keyword{timeManip}

