% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYY2_chr}
\alias{YYYY2_chr}
\title{YYYY2_chr}
\usage{
YYYY2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYY2_chr
}
\examples{
\dontrun{
res <- YYYY2_chr()
}
}
\keyword{timeManip}
