% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{YYYY_chr}
\alias{YYYY_chr}
\title{YYYY_chr}
\usage{
YYYY_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYY_chr
}
\examples{
\dontrun{
res <- YYYY_chr()
}
}
\keyword{timeManip}
