% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYY_m}
\alias{YYYY_m}
\title{Get YYYY_m}
\usage{
YYYY_m(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year as a matrix
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYY_m(res)
}
\keyword{timeManip}

