% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmm}
\alias{YYYYmm}
\title{Get YYYYmm}
\usage{
YYYYmm(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year and month as character
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYYmm(res)
}
\keyword{timeManip}

