% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmm2_chr}
\alias{YYYYmm2_chr}
\title{YYYYmm2_chr}
\usage{
YYYYmm2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmm2_chr
}
\examples{
\dontrun{
res <- YYYYmm2_chr()
}
}
\keyword{timeManip}
