% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{YYYYmm_chr}
\alias{YYYYmm_chr}
\title{YYYYmm_chr}
\usage{
YYYYmm_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmm_chr
}
\examples{
\dontrun{
res <- YYYYmm_chr()
}
}
\keyword{timeManip}
