% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmm_m}
\alias{YYYYmm_m}
\title{Get YYYYmm_m}
\usage{
YYYYmm_m(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year and month as a matrix
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYYmm_m(res)
}
\keyword{timeManip}

