% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmdd}
\alias{YYYYmmdd}
\title{Get YYYYmmdd}
\usage{
YYYYmmdd(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month and day as character
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYYmmdd(res)
}
\keyword{timeManip}
