% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{YYYYmmdd2_chr}
\alias{YYYYmmdd2_chr}
\title{YYYYmmdd2_chr}
\usage{
YYYYmmdd2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmdd2_chr
}
\examples{
\dontrun{
res <- YYYYmmdd2_chr()
}
}
\keyword{timeManip}
