% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmdd_chr}
\alias{YYYYmmdd_chr}
\title{YYYYmmdd_chr}
\usage{
YYYYmmdd_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmdd_chr
}
\examples{
\dontrun{
res <- YYYYmmdd_chr()
}
}
\keyword{timeManip}
