% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmdd_m}
\alias{YYYYmmdd_m}
\title{Get YYYYmmdd_m}
\usage{
YYYYmmdd_m(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month and day as a matrix
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYYmmdd_m(res)
}
\keyword{timeManip}

