% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmddHH}
\alias{YYYYmmddHH}
\title{Get YYYYmmddHH}
\usage{
YYYYmmddHH(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month, day and hour as character
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
YYYYmmddHH(res)
}
\keyword{timeManip}

