% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmddHH2_chr}
\alias{YYYYmmddHH2_chr}
\title{YYYYmmddHH2_chr}
\usage{
YYYYmmddHH2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHH2_chr
}
\examples{
\dontrun{
res <- YYYYmmddHH2_chr()
}
}
\keyword{timeManip}
