% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmddHH_chr}
\alias{YYYYmmddHH_chr}
\title{YYYYmmddHH_chr}
\usage{
YYYYmmddHH_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHH_chr
}
\examples{
\dontrun{
res <- YYYYmmddHH_chr()
}
}
\keyword{timeManip}
