% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmddHHMM}
\alias{YYYYmmddHHMM}
\title{Get YYYYmmddHHMM}
\usage{
YYYYmmddHHMM(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month, day, hour and minute as character
}
\examples{
res <- YYYYmmddHHMM_chr("201501010010")
YYYYmmddHHMM(res)
}
\keyword{timeManip}
