% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmddHHMM2_chr}
\alias{YYYYmmddHHMM2_chr}
\title{YYYYmmddHHMM2_chr}
\usage{
YYYYmmddHHMM2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHHMM2_chr
}
\examples{
\dontrun{
res <- YYYYmmddHHMM2_chr()
}
}
\keyword{timeManip}
