% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmddHHMM_chr}
\alias{YYYYmmddHHMM_chr}
\title{YYYYmmddHHMM_chr}
\usage{
YYYYmmddHHMM_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHHMM_chr
}
\examples{
\dontrun{
res <- YYYYmmddHHMM_chr()
}
}
\keyword{timeManip}

