% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmddHHMM_m}
\alias{YYYYmmddHHMM_m}
\title{Get YYYYmmddHHMM_m}
\usage{
YYYYmmddHHMM_m(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month, day, hour and minute as a matrix
}
\examples{
res <- YYYYmmddHHMM_chr("201501010010")
YYYYmmddHHMM_m(res)
}
\keyword{timeManip}

