% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert.R
\name{YYYYmmddHHMMSS}
\alias{YYYYmmddHHMMSS}
\title{Get YYYYmmddHHMMSS}
\usage{
YYYYmmddHHMMSS(date)
}
\arguments{
\item{date}{POSIXlt}
}
\description{
Get the year, month, day, hour, minute and second as character
}
\examples{
res <- YYYYmmddHHMMSS_chr("20150101001034")
YYYYmmddHHMMSS(res)
}
\keyword{timeManip}

