% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_make.R
\name{YYYYmmddHHMMSS2_chr}
\alias{YYYYmmddHHMMSS2_chr}
\title{YYYYmmddHHMMSS2_chr}
\usage{
YYYYmmddHHMMSS2_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHHMMSS2_chr
}
\examples{
\dontrun{
res <- YYYYmmddHHMMSS2_chr()
}
}
\keyword{timeManip}
