% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{YYYYmmddHHMMSS_chr}
\alias{YYYYmmddHHMMSS_chr}
\title{YYYYmmddHHMMSS_chr}
\usage{
YYYYmmddHHMMSS_chr(date_chr)
}
\arguments{
\item{date_chr}{date in character}
}
\value{
The output is a POSIXlt value
}
\description{
YYYYmmddHHMMSS_chr
}
\examples{
\dontrun{
res <- YYYYmmddHHMMSS_chr()
}
}
\keyword{timeManip}
