% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{addition}
\alias{addition}
\title{Addition of a value to a date}
\usage{
addition(date, timeResolution, v)
}
\arguments{
\item{date}{string of character yyyymmdd}

\item{timeResolution}{choice between "daily","hourly","minute","second"}

\item{v}{positive (resp.negative) real value to be added (resp. substracted).}
}
\description{
The function \code{addition} enables the addition of a value to a date.
The date has no format, nor any time zone.
}
\examples{
res <- addition("20150728","hourly",14)
}
\keyword{timeManip}

