% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contain.R
\name{contain}
\alias{contain}
\title{contain}
\usage{
contain(a, b)
}
\arguments{
\item{a}{POSIXlt timeserie}

\item{b}{POSIXlt timeserie}
}
\description{
contain
}
\examples{
\dontrun{
contain()
}
}
\keyword{timeManip}

