% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{extract}
\usage{
extract(date, indice = NULL)
}
\arguments{
\item{date}{timeManip object}

\item{indice}{indices}
}
\description{
extract
}
\examples{
a <- timeManip(fromPeriod="2013060205",toPeriod="2013062805",
               timeResolution="hourly",precision="hourly")
res <- extract(a,indice=list(offset=3,count=5))
res$fromPeriod()
res$toPeriod()
}
\keyword{timeManip}

