% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{getP}
\alias{getP}
\title{Get the part of the POSIXlt date you want}
\usage{
getP(date, part, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{part}{"year" "month", "day", "hour", "minute", "second"}

\item{CHAR}{boolean as character}
}
\description{
Get the part of the POSIXlt date you want
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
getP(res,"month")
}
\keyword{timeManip}

