% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{hour}
\alias{hour}
\title{Get the hour}
\usage{
hour(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the hour
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
hour(res)
}
\keyword{timeManip}
