% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{indice_subdate}
\alias{indice_subdate}
\title{indice_subdate}
\usage{
indice_subdate(date, d)
}
\arguments{
\item{date}{timeManip object}

\item{d}{sub timeManip object}
}
\description{
indice_subdate
}
\examples{
\dontrun{
indice_subdate()
}
}
\keyword{timeManip}

