% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insec.R
\name{insec}
\alias{insec}
\title{In second}
\usage{
insec(timeResolution, v = 1, INV = FALSE)
}
\arguments{
\item{timeResolution}{choice between "daily","hourly","minute","second"}

\item{v}{value}

\item{INV}{boolean if into non-second}
}
\description{
The function \code{insec()} gives the values in second
}
\examples{
res <- insec(timeResolution="hourly",v=4)
res <- insec(timeResolution="hourly",v=4,INV=TRUE)
}
\keyword{timeManip}

