% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_split.R
\name{second}
\alias{second}
\title{Get the second}
\usage{
second(date, CHAR = FALSE)
}
\arguments{
\item{date}{POSIXlt}

\item{CHAR}{boolean as character}
}
\description{
Get the second
}
\examples{
res <- YYYYmmddHH_chr("2015010100")
second(res)
}
\keyword{timeManip}

