% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManip.R
\name{timeManip}
\alias{timeManip}
\title{timeManip}
\usage{
timeManip(fromPeriod, toPeriod, timeResolution)
}
\arguments{
\item{fromPeriod}{dateChr "YYYYmmddHH..."}

\item{toPeriod}{dateChr "YYYYmmddHH..."}

\item{timeResolution}{choice between "daily","hourly","minute","second",...}
}
\description{
timeManip
}
\examples{
timeManip(fromPeriod="2013060205",toPeriod="2013060205",timeResolution="hourly")

}
\keyword{timeManip}

