% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManip.R
\name{timeManip}
\alias{timeManip}
\title{timeManip}
\usage{
timeManip(fromPeriod, toPeriod, timeResolution, precision = NULL)
}
\arguments{
\item{fromPeriod}{dateChr "YYYYmmddHH..."}

\item{toPeriod}{dateChr "YYYYmmddHH..."}

\item{timeResolution}{"daily","hourly","minute","second",...}

\item{precision}{"daily","hourly","minute","second",...}
}
\description{
timeManip
}
\examples{
res <- timeManip(fromPeriod="2013060205",toPeriod="2013060205",
                 timeResolution="hourly",precision="hourly")
res$seqPeriod()
}
\keyword{timeManip}

