% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManip.R
\name{timeManip}
\alias{timeManip}
\title{timeManip}
\usage{
timeManip(fromPeriod, toPeriod, timeResolution, precision,
  Timeresinsec = NULL, nbStep = NULL, seqPeriod = NULL)
}
\arguments{
\item{fromPeriod}{dateChr "YYYYmmddHH..."}

\item{toPeriod}{dateChr "YYYYmmddHH..."}

\item{timeResolution}{"daily","hourly","minute","second",...}

\item{precision}{"daily","hourly","minute","second",...}

\item{Timeresinsec}{integer}

\item{nbStep}{integer}

\item{seqPeriod}{POSIXct timeserie}
}
\description{
timeManip
}
\examples{
timeManip(fromPeriod="2013060205",toPeriod="2013060205",timeResolution="hourly",precision="hourly")

}
\keyword{timeManip}

